requirements.txt
yfinance==0.2.64
matplotlib
streamlit
pandas
openai
beautifulsoup4
requests
PyPDF2
python-dotenv
plotly
fastapi
uvicorn
